web-ui/requirements.txt
# requirements.txt
Flask==3.0.3
gunicorn==22.0.0
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
requests==2.32.3
pyyaml==6.0.2
ansible-runner==2.4.0